clean: move statisticCache class into a new file
